app/requirements.txt
uvicorn==0.15.0
fastapi==0.70.0
pydantic==1.8.2
requests==2.25.1
scikit-learn==1.0.2
ipykernel~=6.16.2
pandas~=1.3.5
scipy==1.10.1
xgboost==1.7.4
python-dotenv==1.0.0
